fix : LocalPlace 등록 시, contentTypeId 도 같이 등록되게 변경 및 parking, businessHour

+++ src/main/java/com/example/ieumapi/localplace/domain/LocalPlace.java
@@ -47,7 +47,7 @@ public class LocalPlace {
     private String ktoContentId; // 한국관광공사 contentId, nullable
 
     @Column(nullable = false)
-    private Long contentTypeId;
+    private int contentTypeId;
 
     @Enumerated(EnumType.STRING)
     @Column(nullable = false)
@@ -68,7 +68,7 @@ public class LocalPlace {
         String userNickName,
         Source source,
         String ktoContentId,
-        Long contentTypeId,
+        int contentTypeId,
         PlaceCategory category,
         String parking,
         String businessHours){
